* commit '6e271c19c4f7404eea86c765652724e87bd30183':
  fix delete bug

+++ dataset-service/src/main/java/org/eea/dataset/service/impl/DatasetServiceImpl.java
@@ -883,8 +883,10 @@ public class DatasetServiceImpl implements DatasetService {
         }
       }
       // delete all fks
-      LOG.info("records with fk's to delete {}", recordsToDelete);
-      recordRepository.deleteRecordWithIdIn(recordsToDelete);
+      if (!recordsToDelete.isEmpty()) {
+        LOG.info("records with fk's to delete {}", recordsToDelete);
+        recordRepository.deleteRecordWithIdIn(recordsToDelete);
+      }
     }
   }
